Added an actual webpage.
Updated the WSGI file so that it actually loads a page.
Should be portable, since it uses the provided document root as the
base, which ought to allow it to run anywhere the project is set up in
the same way as here.
Also included a .gitignore file, currently only set to exclude my
Komodo project file and any autosave files.