AJAX Example
AJAX now works. At some point some of this stuff might need to be split
up, but for now it’s probably all good. Server log included, and not
added to .gitignore in case we ever need the records. In the future
I’ll need to figure out how to truncate it if it gets too long.