Now With 100% More CSS!
CSS added. Trump spins now, because that is a Good Idea. Had to update
the app to change content-type to “text/CSS” when sending back a CSS
file, because apparently that’s important. my_app.wsgi continues its
path towards a refactor, but at this point it supports just about
everything we’ll need.